Refatorando projeto

+++ java-clientes/src/main/java/br/com/cotiinformatica/services/EnderecoService.java
@@ -3,6 +3,7 @@ package br.com.cotiinformatica.services;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import javax.transaction.Transactional;
 
@@ -10,6 +11,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import br.com.cotiinformatica.dtos.EnderecoDTO;
+import br.com.cotiinformatica.entities.Cliente;
 import br.com.cotiinformatica.entities.Endereco;
 import br.com.cotiinformatica.exceptions.EntityNotFoundException;
 import br.com.cotiinformatica.repositories.EnderecoRepository;
@@ -26,8 +28,8 @@ public class EnderecoService {
 	public Endereco cadastrar(EnderecoDTO dto) {
 
 		// buscando um endereço existente no banco
-		Optional<Endereco> result = enderecoRepository.findByNumeroAndCepAndComplemento(
-				dto.getNumero(), dto.getCep(), dto.getComplemento());
+		Optional<Endereco> result = enderecoRepository.findByLogradouroAndNumeroAndComplementoAndCep(
+				dto.getLogradouro(), dto.getNumero(), dto.getComplemento(), dto.getCep());
 
 		// caso exista, retorne o mesmo endereço
 		if (result.isPresent()) {
@@ -68,26 +70,34 @@ public class EnderecoService {
 	}
 
 	public Endereco atualizar(EnderecoDTO dto) {
-
-		// buscando um endereço existente no banco
-		Optional<Endereco> result = enderecoRepository.findByNumeroAndCepAndComplemento(
-				dto.getNumero(), dto.getCep(), dto.getComplemento());
 		
-		// caso exista, retorne o mesmo endereço
-		if(result.isPresent()) {
-			Endereco endereco = result.get();
+		// buscando um endereço existente no banco
+		Optional<Endereco> result = enderecoRepository.findById(dto.getIdEndereco());
+		
+		// Endereço encontrado
+		Endereco endereco = result.get();
+	
+		// recebendo lista de clientes do endereço
+		Set<Cliente> lista = endereco.getClientes();
+
+		if (lista.size() > 1) {
+			// se o mesmo endereço pertencer a outro cliente
+			Endereco end = mapper.map(dto, Endereco.class);
+			end.setIdEndereco(null);
+			enderecoRepository.save(end);
+			return end;
+		}
+		
+		// endereco do dto
+		Endereco end2 = dtoToEndereco(dto);
+		if(endereco.equals(end2)){
 			return endereco;
 		}
-		
-		Optional<Endereco> result2 = enderecoRepository.findById(dto.getIdEndereco());
-		
-		if (result2.isEmpty()) {
-			throw new EntityNotFoundException("Endereço não encontrado.");
-		}
-
-		Endereco endereco = result2.get();
+		 
+		// convertendo dto em Endereco
 		endereco = dtoToEndereco(dto);
 
+		// salvando endereço atualizado
 		enderecoRepository.save(endereco);
 
 		return endereco;
@@ -102,15 +112,23 @@ public class EnderecoService {
 		}
 
 		Endereco endereco = result.get();
+		
+		// recebendo lista de clientes do endereço
+		Set<Cliente> lista = endereco.getClientes();
+		
+		// não deletar caso o endereço pertença a mais de um Cliente
+		if(lista.size() > 1) {
+			return null;
+		}
 
 		enderecoRepository.delete(endereco);
 
 		return "Endereço excluído com sucesso.";
 	}
 
-// Método para converter um EnderecoDTO em Endereco
-	public Endereco dtoToEndereco(EnderecoDTO enderecoDto) {
-		Endereco endereco = mapper.map(enderecoDto, Endereco.class);
+    // Método para converter um EnderecoDTO em Endereco
+	public Endereco dtoToEndereco(EnderecoDTO dto) {
+		Endereco endereco = mapper.map(dto, Endereco.class);
 		return endereco;
 	}
